Rename derEncoding.c/h to schnorr.c/h
