+++ src/schnorr.c
@@ -1,4 +1,4 @@
-#include "derEncoding.h"
+#include "schnorr.h"
 
 /* ----------------------------------------------------------------------- */
 /*                                                                         */
