+++ src/schnorr.h
